backend-working/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
pytesseract>=0.3.8
PyMuPDF>=1.18.0
spacy>=3.0.0
transformers>=4.5.0
torch>=1.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz